Polyakov Alexander/telegram_bot/api/requirements.txt
Flask==2.3.3
gunicorn==21.2.0
requests==2.31.0
numpy==1.24.3
Pillow==10.0.0
tensorflow==2.13.0
scikit-learn==1.2.2
ultralytics==8.0.184
pytesseract==0.3.10
python-Levenshtein==0.21.1